done makinng crud of usre and place

+++ controllers/usersController.mjs
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import User from "../models/User.mjs";
 import Place from "../models/Place.mjs";
 import HttpError from "../models/httpError.mjs";
@@ -20,11 +21,36 @@ export const createUser = async (req, res, next) => {
   }
 };
 
+// UPDATE SINGLE USER
+export const updateUser = async (req, res, next) => {
+  const newData = req.body;
+  const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: "Invalid user id" });
+  }
+
+  try {
+    const myExistPlace = await User.findById(id);
+    if (!myExistPlace) {
+      return res.status(404).json({ message: "user not found" });
+    }
+    const updatedUser = await User.findByIdAndUpdate(id, newData, {
+      new: true,
+    });
+    return res.status(200).json({ message: "Got you User", updatedUser });
+  } catch (error) {
+    return res.status(500).json({ success: false, error: error.message });
+  }
+};
+
 // GET SINGLE USER
 export const getUser = async (req, res, next) => {
   const { id } = req.params; // ya req.body.id, jahan se id aa rahi ho
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: "Invalid user id" });
+  }
   try {
-    const user = await User.findById(id).populate("places");
+    const user = await User.findById(id).populate("placesInUser"); // populate id se pura place fetch kar lega
     if (!user) {
       return res.status(404).json({ error: "User not found." });
     }
@@ -38,15 +64,19 @@ export const getUser = async (req, res, next) => {
 // GET ALL USERS
 export const getUsers = async (req, res, next) => {
   try {
-    const allUsers = await User.find().populate("places");
+    const allUsers = await User.find().populate("placesInUser");
     res.status(200).json({ allUsers });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 };
 
+// DELETE SINGLE USER
 export const deleteUser = async (req, res, next) => {
   const userId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(userId)) {
+    return res.status(400).json({ message: "Invalid user id" });
+  }
   try {
     const vanishUser = await User.findByIdAndDelete(userId);
     res.status(200).json({ message: "deleted", vanishUser });
@@ -55,3 +85,10 @@ export const deleteUser = async (req, res, next) => {
   }
 };
 
+// DELETE ALL USER
+export const deleteAllUsers = async (req, res, next) => {
+  await User.deleteMany();
+  return res
+    .status(200)
+    .json({ message: "all Users are gone. Sorry about that!" });
+};
